Restructuring Code (WIP)
Writing everything in two big elif blocks turned out to be a bad decision. Some changes will need to be made to fix bugs caused by this. This commit simply starts this by making the first block of elifs into seperate ifs. Much more needs to be done to fix the issue.